# life-xp-activity-service-api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Activity Service API
  description: API для управления активностями и их коэффициентами XP
  version: 1.0.0

tags:
  - name: ActivityCommand
    description: API для команд активностей (создание, обновление, удаление)
  - name: ActivityQuery
    description: API для запросов активностей (получение информации)
  - name: UserActivityCommand
    description: API для команд активностей юзера(создание, обновление, удаление)
  - name: UserActivityQuery
    description: API для запросов активностей юзера (получение информации)

paths:
  /activities:
    get:
      summary: Получить список всех активностей
      description: Возвращает все доступные активности
      operationId: getActivities
      tags:
        - ActivityQuery
      responses:
        "200":
          description: Успешный ответ с массивом активностей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityResponse"
  /activities/create:
    post:
      summary: Создать новую активность
      description: Добавляет новую активность
      operationId: createActivity
      tags:
        - ActivityCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "201":
          description: Активность создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateActivityResponse"

  /activities/{id}:
    get:
      summary: Получить активность по ID
      description: Возвращает данные об активности
      operationId: getActivityById
      tags:
        - ActivityQuery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешный ответ с активностью
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
        "404":
          description: Активность не найдена
    put:
      summary: Обновить активность
      description: Обновляет данные об активности
      operationId: updateActivity
      tags:
        - ActivityCommand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "200":
          description: Активность обновлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
    delete:
      summary: Удалить активность
      description: Удаляет активность по ID
      operationId: deleteActivity
      tags:
        - ActivityCommand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Активность удалена
  /user-activities/{userId}:
    get:
      summary: Получить активности пользователя
      description: Возвращает список активностей, к которым привязан пользователь
      operationId: getUserActivities
      tags:
        - UserActivityQuery
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешный ответ с массивом активностей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserActivityResponse"
  /user-activities/create:
    post:
      summary: Добавить активность пользователю
      description: Создает новую запись активности для пользователя
      operationId: addUserActivity
      tags:
        - UserActivityCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserActivityRequest"
      responses:
        "200":
          description: Активность успешно добавлена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserActivityResponse"
        "400":
          description: Неверные входные данные

components:
  schemas:
    ActivityResponse:
      type: object
      properties:
        name:
          type: string
          example: coding
        description:
          type: string
          example: Программирование
        category:
          $ref: '#/components/schemas/ActivityCategory'
        unit:
          $ref: '#/components/schemas/ActivityUnit'
        baseXp:
          type: integer
          format: int32
          description: Базовый опыт за единицу активности
    CreateActivityRequest:
      type: object
      required:
        - name
        - description
        - category
        - unit
        - baseXp
      properties:
        name:
          type: string
          example: coding
        description:
          type: string
          example: Программирование
        category:
          $ref: '#/components/schemas/ActivityCategory'
        unit:
          $ref: '#/components/schemas/ActivityUnit'
        baseXp:
          type: number
          format: double
          description: Базовый опыт за единицу активности

    CreateActivityResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid

    MultiplierResponse:
      type: object
      properties:
        activityId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        xpMultiplier:
          type: number
          format: float
          example: 1.7

    MultiplierInput:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        xpMultiplier:
          type: number
          format: float
          example: 1.7

    CreateUserActivityResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid

    CreateUserActivityRequest:
      type: object
      required:
        - userId
        - activityId
      properties:
        userId:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        customXp:
          type: integer
          description: Пользовательское значение XP

    UserActivityResponse:
      type: object
      required:
        - activityId
        - activityName
        - category
        - unit
        - baseXp
      properties:
        activityId:
          type: string
          format: uuid
        activityName:
          type: string
          example: coding
        category:
          $ref: '#/components/schemas/ActivityCategory'
        unit:
          $ref: '#/components/schemas/ActivityUnit'
        baseXp:
          type: integer
          format: int32
        customXp:
          type: integer
          description: Пользовательское значение XP

    ActivityUnit:
      type: string
      enum:
        - HOURS
        - KILOMETERS
        - SESSIONS
      description: Единица измерения активности

    ActivityCategory:
      type: string
      enum:
        - SPORT
        - EDUCATION
        - HEALTH
        - ENTERTAINMENT
        - OTHER
      description: Категория активности